move StringUtils to Shared folder
